requirements.txt
streamlit==1.24.0
Pillow==9.5.0